api/requirements.txt
# FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Config / validation
pydantic==2.9.0
pydantic-settings==2.5.2

# Google Gemini (embeddings)
google-generativeai==0.8.3

# OpenAI (LLM — GPT-4o)
openai==1.51.0

# Pinecone (production vector store)
pinecone==5.3.0

# Chroma (dev/local vector store)
chromadb==0.5.18

# Document parsing
pymupdf==1.24.10        # PDF: PyMuPDF (fitz) — reliable page extraction
python-docx==1.1.2      # DOCX: section heading extraction

# Chunking
langchain-text-splitters==0.3.0

# Async HTTP (used for bounded Gemini calls if needed)
httpx==0.27.2
